+++ CHANGELOG.md
@@ -2,6 +2,8 @@ v3.1.7
 * Fix drop at incorrect place when player is dead.
 * Fix skillPickup's incorrect depth.
 * Add enforcerZ2Reload to the blacklist.
+* Fix huntressX2 can't freeze correctly.
+* Fix drop incorrect skill sometimes.
 
 v3.1.6
 * Add valid check to SkillModifierData.